Change source code compliance level to 1.7

+++ lenox/src/main/java/es/inteco/crawler/sexista/core/dto/impl/BasePagingDtoImpl.java
@@ -17,7 +17,7 @@ public class BasePagingDtoImpl extends BaseDto implements BasePagingDto {
      * Tamaño de página por defecto.
      */
     private static final int DEFAULT_PAGE_SIZE = Integer.valueOf(ConfigUtil.getConfiguracion()
-            .getProperty("maxRows")).intValue();
+            .getProperty("maxRows"));
 
     /**
      * Registro de inicio.
